src/test/java/reservationapp/PublicRoomsControllerTest.java
package reservationapp;

import com.fasterxml.jackson.core.type.TypeReference;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateDeserializer;
import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateSerializer;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import reservationapp.api.ReservationRequest;
import reservationapp.repository.RoomRecord;
import reservationapp.repository.RoomRepository;

import java.math.BigDecimal;
import java.time.LocalDate;
import java.time.format.DateTimeFormatter;
import java.util.Collections;
import java.util.List;

import static com.fasterxml.jackson.databind.SerializationFeature.WRITE_DATES_AS_TIMESTAMPS;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.springframework.http.MediaType.APPLICATION_JSON;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@RunWith(SpringRunner.class)
@WebMvcTest(PublicRoomsController.class)
public class PublicRoomsControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    ReservationService reservationService;

    @MockBean
    RoomRepository roomRepository;

    private static final ObjectMapper MAPPER = new ObjectMapper();

    static {
        Jackson2ObjectMapperBuilder builder = new Jackson2ObjectMapperBuilder();

        JavaTimeModule javaTimeModule = new JavaTimeModule();

        javaTimeModule.addDeserializer(
                LocalDate.class,
                new LocalDateDeserializer(DateTimeFormatter.ofPattern("yyyy-MM-dd"))
        );

        javaTimeModule.addSerializer(
                LocalDate.class,
                new LocalDateSerializer(DateTimeFormatter.ofPattern("yyyy-MM-dd"))
        );

        builder.modules(javaTimeModule);
        builder.featuresToDisable(WRITE_DATES_AS_TIMESTAMPS);

        MAPPER.registerModule(javaTimeModule);
    }

    @Test
    public void should_be_able_to_reserve_room() throws Exception {

        ReservationRequest reservationRequest = new ReservationRequest(
                LocalDate.now(),
                LocalDate.now().plusDays(3)
        );

        String json = MAPPER.writeValueAsString(reservationRequest);

        mockMvc.perform(
                post("/public-api/rooms/{room-number}/reserve", 420)
                        .content(json)
                        .contentType(APPLICATION_JSON)
                        .accept(APPLICATION_JSON))
                .andExpect(status().isOk());

    }

    @Test
    public void should_be_able_to_see_available_rooms() throws Exception {

        RoomRecord roomRecord = new RoomRecord(
                420,
                new BigDecimal("100.0"),
                5
        );

        Mockito.lenient()
                .when(reservationService.availableRooms(any(), any()))
                .thenReturn(Collections.singletonList(roomRecord));

        String response = mockMvc.perform(
                get("/public-api/rooms/search")
                        .param("checkIn", "2019-01-02")
                        .param("checkOut", "2019-01-08"))
                .andExpect(status().isOk())
                .andReturn()
                .getResponse()
                .getContentAsString();

        List<RoomRecord> result = MAPPER.readValue(response, new TypeReference<List<RoomRecord>>() {
        });

        assertEquals(Collections.singletonList(roomRecord), result);

    }
}